SCOP FINAL

+++ Projetos QT/SCOP/mainwindow.cpp
@@ -1,8 +1,6 @@
 #include "mainwindow.h"
 #include "ui_mainwindow.h"
 
-
-
 Cadastro temp;
 Objeto a;
 
@@ -11,6 +9,7 @@ MainWindow::MainWindow(QWidget *parent)
     , ui(new Ui::MainWindow)
 {
     ui->setupUi(this);
+
 }
 
 MainWindow::~MainWindow()
@@ -29,30 +28,27 @@ void MainWindow::on_btncadastrar_clicked()
         temp.setEstado(ui->estadoatual->currentText());
         temp.setDestino(ui->destinoObj->text());
         temp.setPeso(ui->pesoobj->text().toDouble());
-
         int quantidade_linhas = ui->tabela->rowCount();
-
         if(a.objetoExiste(temp) == 1){
-            QMessageBox::critical(this,"Opa...","O objeto já existe, favor editar na aba da tabela");
+            QMessageBox::critical(this,"Erro","O objeto já existe, favor inserir um objeto diferente");
         }else{
             a.inserirObjeto(temp);
             ui->tabela->insertRow(quantidade_linhas);
             inserirNaTabela(temp,quantidade_linhas);
-            QMessageBox::information(this,"Cadastro:","Cadastrado com sucesso!");
+            QMessageBox::information(this,"Cadastro","Cadastrado com sucesso!");
             ui->tabela->setColumnWidth(0,195);
             ui->tabela->setColumnWidth(3,199);
             ui->tabela->setColumnWidth(4,250);
             atualizarEstatisticas();
+            ui->inputCi->clear();
+            ui->inputObjeto->clear();
+            ui->destinoObj->clear();
+            ui->valorobj->clear();
+            ui->pesoobj->clear();
         }
     }else{
         QMessageBox::critical(this,"Erro","Um dos parâmetros estão vazios, favor preencher e tentar novamente.");
     }
-    ui->inputCi->clear();
-    ui->inputObjeto->clear();
-    ui->destinoObj->clear();
-    ui->valorobj->clear();
-    ui->pesoobj->clear();
-
 }
 
 void MainWindow::inserirNaTabela(Cadastro a, int linha)
@@ -110,6 +106,9 @@ void MainWindow::on_actionAbrir_triggered()
             inserirNaTabela(a[i],i);
         }
         atualizarEstatisticas();
+        ui->tabela->setColumnWidth(0,195);
+        ui->tabela->setColumnWidth(3,199);
+        ui->tabela->setColumnWidth(4,250);
         QMessageBox::information(this,"Arquivo"," O arquivo foi lido, cheque a tabela!");
     }
 }
@@ -129,7 +128,7 @@ void MainWindow::on_actionSair_triggered()
     QApplication::quit();
 }
 
-void MainWindow::on_tabela_cellDoubleClicked(int row, int column)
+void MainWindow::on_tabela_cellClicked(int row, int column)
 {
     if(column == 0){
         QMessageBox::StandardButton resp = QMessageBox::question(this, "Editar Itens", "Você deseja editar este item?");
@@ -145,9 +144,13 @@ void MainWindow::on_tabela_cellDoubleClicked(int row, int column)
                 temp.setDestino(a[row].getDestino());
                 temp.setPeso(a[row].getPeso());
                 a.eraseObj(txt);
-                a.inserirObjeto(temp);
-                ui->tabela->setItem(row, column, new QTableWidgetItem(temp.getObj()));
-                atualizarEstatisticas();
+                if(a.objetoExiste(temp) == 0){
+                    a.inserirObjeto(temp);
+                    ui->tabela->setItem(row, column, new QTableWidgetItem(temp.getObj()));
+                    atualizarEstatisticas();
+                }else
+                    QMessageBox::critical(this,"Erro","O Objeto a ser editado já existe!");
+
             }else{
                 QMessageBox::critical(this, "Erro", "O objeto a ser editado está vazio.");
             }
@@ -159,7 +162,6 @@ void MainWindow::on_tabela_cellDoubleClicked(int row, int column)
             bool ok;
             QString txt = QInputDialog::getText(this, "Alterar CI", "Digite o novo CI", QLineEdit::Normal,"",&ok);
             if(ok and !txt.isEmpty()){
-
                 temp.setObj(a[row].getObj());
                 temp.setCi(txt);
                 temp.setData(a[row].getData());
@@ -182,7 +184,6 @@ void MainWindow::on_tabela_cellDoubleClicked(int row, int column)
             bool ok;
             QString txt = QInputDialog::getText(this, "Alterar Valor", "Digite o novo valor", QLineEdit::Normal,"",&ok);
             if(ok and !txt.isEmpty()){
-
                 temp.setObj(a[row].getObj());
                 temp.setCi(a[row].getCi());
                 temp.setData(a[row].getData());
@@ -250,7 +251,6 @@ void MainWindow::on_tabela_cellDoubleClicked(int row, int column)
             bool ok;
             QString txt = QInputDialog::getText(this, "Alterar Destino", "Digite o novo destino", QLineEdit::Normal,"",&ok);
             if(ok and !txt.isEmpty()){
-
                 temp.setObj(a[row].getObj());
                 temp.setCi(a[row].getCi());
                 temp.setData(a[row].getData());
@@ -319,13 +319,15 @@ void MainWindow::on_actionExcluir_triggered()
 void MainWindow::on_actionLimpar_Tabela_triggered()
 {
     int qtdeLinhas = ui->tabela->rowCount();
-    for(int i = 0; i < qtdeLinhas ; i++)
-        ui->tabela->removeRow(0);
+    for(int i = 0; i < qtdeLinhas ; i++)ui->tabela->removeRow(0);
+
 }
 
 void MainWindow::on_limpartab_clicked()
 {
     on_actionLimpar_Tabela_triggered();
+    a.limparTudo();
+    atualizarEstatisticas();
     QMessageBox::information(this,"Limpo","A tabela foi resetada com sucesso");
 }
